requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
PyYAML==6.0.1
